Update glider_trajectory_uv_template_v.0.0.cdl
Changed 'time_avg' variable name to 'time_uv' and all references (attributes) to the variable. Changed 'slocum' variable to 'platform' and all references (attributes) to the variable.  Added the 'glider_type' attribute to the 'platform' variable in order to determine the type of glider represented in the file.  Added global "platform" attribute according to NODC trajectory template.